requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
bzip2=1.0.8=h93a5062_5
ca-certificates=2024.2.2=hf0a4a13_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
charset-normalizer=3.3.2=pypi_0
google-api-core=2.19.0=pypi_0
google-api-python-client=2.129.0=pypi_0
google-auth=2.29.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-auth-oauthlib=1.2.0=pypi_0
googleapis-common-protos=1.63.0=pypi_0
httplib2=0.22.0=pypi_0
idna=3.7=pypi_0
libexpat=2.6.2=hebf3989_0
libffi=3.4.2=h3422bc3_5
libsqlite=3.45.3=h091b4b1_0
libzlib=1.2.13=h53f4e23_5
ncurses=6.5=hb89a1cb_0
oauthlib=3.2.2=pypi_0
openssl=3.3.0=hfb2fe0b_2
pip=24.0=pyhd8ed1ab_0
proto-plus=1.23.0=pypi_0
protobuf=4.25.3=pypi_0
pyactiveresource=2.2.2=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pyjwt=2.8.0=pypi_0
pyparsing=3.1.2=pypi_0
python=3.12.3=h4a7b5fc_0_cpython
python-dotenv=1.0.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h92ec313_1
requests=2.32.2=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9=pypi_0
setuptools=69.5.1=pyhd8ed1ab_0
shopifyapi=12.5.0=pypi_0
six=1.16.0=pypi_0
tk=8.6.13=h5083fa2_1
tzdata=2024a=h0c530f3_0
uritemplate=4.1.1=pypi_0
urllib3=2.2.1=pypi_0
wheel=0.43.0=pyhd8ed1ab_1
xz=5.2.6=h57fd34a_0
